+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.9.0]
+## Features
+- Added support for GIA (unsorted and sorted)
+- Added support for bedtk (unsorted and sorted)
+- Added support for igd
+- Sort output of bench into label subfolders
+
 # [0.8.0]
 ## Features
 - Added support for granges (https://github.com/vsbuffalo/granges)